pyproject.toml
[project]
name = "uv-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.0",
    "google-generativeai>=0.8.6",
    "langchain>=1.2.9",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.0",
    "numpy>=2.4.2",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "mkdocs>=1.6.1",
    "mkdocs-gen-files>=0.6.0",
    "mkdocs-literate-nav>=0.6.2",
    "mkdocs-material>=9.7.1",
    "mkdocstrings[python]>=1.0.3",
    "mypy>=1.19.1",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.0",
]

[tool.pyright]
pythonVersion = "3.12"
venvPath = "."
venv = ".venv"

[tool.ruff]
line-length = 88
lint.select = ["E", "F", "W", "C90", "B", "I", "D", "ANN"]
exclude = ["__init__.py"]
src = ["."]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**/*" = ["D", "S101", "ANN", "PLR2004"]

[tool.mypy]
python_version = 3.12
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_calls = true
disallow_untyped_defs = true
ignore_missing_imports = true
no_implicit_optional = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_configs = true
warn_unused_ignores = true
exclude = ["docs"]

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_incomplete_defs = false
warn_return_any = false

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --cov"
testpaths = ["tests"]
